Updated script with Makecert
Added support for the name of the subscription as creation of the Automation connection fails if you don't set your context first. 
Also added support for using MakeCert.  As the cmdlets first published with this only work if you are on windows 10 and server 2016.  Set the variable for pathToMakeCert to 'C:\Program Files (x86)\Fiddler2\makecert.exe'  Also added logic to see if the certificate is already in the Localhost certificate store.